Refactor: 컴포넌트명 파스칼 케이스로 변경 및 폴더 구조 변경
