+++ app/(route)/my/reservations/_components/ReviewButton/index.tsx
@@ -1,10 +1,10 @@
+import ReviewModal from '@/(route)/my/reservations/_components/ReviewModal';
+
 import type { Reservations } from '@/_types/myReservations';
 
 import { useModal } from '@/_hooks/useModal';
 
 import Button from '@/_components/button';
-
-import ReviewModal from '../ReviewModal';
 
 interface ReviewButtonProps {
   data: Reservations;
